fix(ci): python-coverage-shards — split remote_repair.py under 1500-line maintainability limit
The first-party file-size guardrail (test_first_party_code_files_stay_under_line_limit)
failed because src/awf/runtime/pr_monitor_runner/remote_repair.py reached 1817 lines.
Move the protected-scope repair operations into a new sibling module
remote_repair_protected.py and re-point the RunnerDelegatesMixin bindings; both files
now sit under the limit. Update the part_008 test to reference the relocated symbols
from their new homes (remote_ops._ProtectedScopePushBlock,
remote_repair_protected.protected_file_diffs_for_committed_paths).
